SEC-2: float ledger → integer smallest-unit (µLENS / µLXC) (#284)
* SEC-2: float ledger → integer smallest-unit (µLENS / µLXC)

Convert every Tier-1 conserved token amount from DOUBLE PRECISION float to
BIGINT integer smallest-unit (1 LENS = 1e6 µLENS; 1 LXC = 1e6 µLXC) so ledger
arithmetic is exact and conservation holds with zero drift. Delete the
roundTo(v,6) band-aid at all call sites + both helper defs; replace the
povi/stakes.go and billing float-epsilon zero-checks with exact integer ==.
House-favoring integer rounding at the 6 meaning-change sites (payouts round
DOWN, charges round UP; remainder retained by the protocol). Emit integer µ-unit
on the JSON boundary with explicit _ulens/_ulxc/_uusd field names. Forward
migrations 0081/0082 ALTER the columns to BIGINT. The SEC-2 drift test flips
RED→GREEN (exact conservation, 0 drift).

STOP before merge — founder review pending. Tier 2/3 (rates/USD) deferred.


* SEC-2 follow-up: migration view-dependency fix + real-PG round-trip proof

- 0081 now DROPs the two royalty MARGIN views (pool_royalty_margin,
  distill_royalty_margin) before ALTERing minted_amount and RECREATEs them after
  (Postgres refuses ALTER TYPE on a column a view reads). The recreated views
  divide minted_amount (now µLENS) by 1e6 so margin_usd = avoided_cogs_usd −
  minted_LENS stays a coherent Tier-3 dollar identity.
- New real-Postgres integration test applies the ACTUAL migration set (0001…0082
  via the real dbmigrate runner over migrations.FS), asserts the physical column
  types are BIGINT, and round-trips LENS credit→debit→balance (=1.0 LENS exact)
  and LXC credit→spend→balance (=2.0 LXC exact) through the real store. Proven
  green against pgvector/pg16. pgxmock could not have caught a BIGINT/DOUBLE
  mismatch — this closes that gap.


* fix(lint): split typed money consts out of mixed const groups (staticcheck SA9004)


* SEC-2: fix real-PG integration tests (float→µLENS/µLXC) + margin analytics

The real-PG integration tests (gated on LENS_TEST_DATABASE_URL, so they never
ran in the local no-DB suite) asserted stale float LENS/LXC values and declared
their own inline DDL as DOUBLE PRECISION. Converted across poolroyalty
(distill/latency/eval/routing/confidential/revoker/cap/holdback/margin), billing
webhook, mining reputation, and povi concurrency: inline DDL money columns →
BIGINT, scan helpers/vars → int64, assertions → µ-unit (micro()). Fixed two real
bugs the float→int compile hid: a RevokeHeldTx(…, 1.0) that became 1 µLENS
(should be micro(1)) and the povi concurrency stake consts that were 1e6× too
small (S=100 µLENS depleted after ~7 halving-slashes).

margin analytics (Tier-3, pulled in by minted_amount→BIGINT): MarginSummaryRow.
MintedLENS → int64 µLENS; the two royalty margin views (0044/0064 recreated in
0081, plus every test's inline copy) now derive margin_usd = avoided_cogs_usd −
minted_amount/1e6 to keep the dollar identity coherent; cmd/lens margin DTO +
handler tests updated.

TestSEC2_RealPG isolated into its own private schema (search_path) so the
shared-DB `go test ./...` run can't collide. Full suite `go test -race -p 1 ./...`
green against pgvector/pg16.


* SEC-2: renumber migrations 0081/0082 → 0082/0083 (rebase onto keel's 0081)

main gained 0080_keel_findings + 0081_keel_findings_mode from the keel merge;
SEC-2's 0081_sec2_lens collided with keel's 0081_keel_findings_mode. Renumber
SEC-2's two money-type migrations to 0082 (LENS) / 0083 (LXC) so versions are
unique and ordered after keel's. Discovered via //go:embed *.sql (no Go
references); from-zero apply only (never deployed).


---------

+++ internal/poolroyalty/detector_integration_test.go
@@ -46,7 +46,7 @@ func detectorTestPool(t *testing.T) *pgxpool.Pool {
 			model                    TEXT NOT NULL DEFAULT '',
 			similarity               DOUBLE PRECISION NOT NULL DEFAULT 0,
 			avoided_cogs_usd         DOUBLE PRECISION NOT NULL DEFAULT 0,
-			minted_amount            DOUBLE PRECISION NOT NULL DEFAULT 0,
+			minted_amount BIGINT NOT NULL DEFAULT 0,
 			answer_sha256            TEXT NOT NULL DEFAULT '',
 			prompt_sha256            TEXT NOT NULL DEFAULT '',
 			status                   TEXT NOT NULL DEFAULT 'final',
